+++ eyes.selenium.java/src/main/java/com/applitools/eyes/selenium/wrappers/EyesTargetLocator.java
@@ -139,8 +139,7 @@ public class EyesTargetLocator implements WebDriver.TargetLocator {
         logger.verbose("enter");
         if (driver.getFrameChain().size() != 0) {
             logger.verbose("Making preparations...");
-            Frame frame = driver.getFrameChain().pop();
-            frame.returnToOriginalPosition(driver);
+            driver.getFrameChain().pop();
             logger.verbose("Done! Switching to parent frame...");
             parentFrame(logger, targetLocator, driver.getFrameChain());
         }