+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [vNext]
 ### Fixed
 - Calling `switchTo` with `EyesWebDriver` now works properly. [Trello 1818](https://trello.com/c/488BZ24S)
+- Fixed a bug with limiting the screenshot size. [Trello 1991](https://trello.com/c/2iCNfoI7)
 ### Updated
 - Screenshot retry mechanism is now more efficient. [Trello 1866](https://trello.com/c/KyxkI6Bu)
 - Supporting visual viewports in UFG. [Trello 1957](https://trello.com/c/jWvdBwex)